ORM delete all a 13

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -19,7 +19,8 @@ if __name__ == "__main__":
     Session = sessionmaker(bind=engine)
     session = Session()
 
-    users = session.query(State).filter(State.name.ilike("%a%")).all()
+    users = session.query(State).all()
     for user in users:
-        session.delete(user)
-        session.commit()
+        if 'a' in user.__dict__['name']:
+            session.delete(user)
+    session.commit()
